Made the nfa -> dfa conversion actually work.
It took several rewrites, and the code is still a bit clunky, but it
successfully produces an accurate dfa.
Also, the way labels are stored was changed, now they are stored
elementwise and with sets so that duplicate labels are not created.
Instead of having a label for the start state, there is now a variable
in DFAs and NFAs which stores the starting state. This is necessary for
the conversion between NFA to DFA to be accurate.