Fixed bugs / imporoved context_free_grammars
Fixed bugs so that it actually correctly generates syntax trees.
Added associativity to productions (left or right).
Added the capability to resolve ambiguous grammars using associativity
to the CFGFA generator.
Added the argument input_transform to cfgfa.test() so that if you have
input that is not in the exact form that the DFA was designed to accept,
you can put in a lambda that will do that, so the DFA can be executed on
the data without having to modify the original data to fit in the DFA.