Fixed bugs in the lexical analysis phase.
Added a new token type, EOF, which is automatically added to the end of
token lists produced by convert_to_tokens. This is required for the
syntax analysis phase.
Fixed a bug where the last token in a file would not be included in the
output of convert_to_tokens.