Added header for Scopes and other objects related to it.
Scopes function as lookup tables and code pages. They can contain
function, variable, and type definitions. They can have a parent scope
set so that when one of these is looked up by name, and the root scope
does not have it, it can look in the parent scope. Scopes additionally
contain a set of Command s to be run. Command objects come in a few
types, the most common being FuncCommand, which calls a FuncScope (a
special scope which has input and output variables specified.) All
builtin functions, including assignment, are represented as emtpy
FuncScope objects.