Created Value class.
This class replaces the roles of Variable, Literal, and FuncInput. It
is an all-purpose value container, usable for constant values as well as
variables. This also makes it usable as a FuncInput replacement, as
FuncInput was previously created to combine the roles of Literal and
Variable. This class will also facilitate future optimization efforts,
as regular variables can be turned in to constants if their value is
determined to be constant.